Add silk touch blocks detection and tool anti-switch
some blocks like glass and sea lanterns were not detected by auto tool, resulting them being broken without silktouch.
This commit adds a modifiable list of blocks that will be mined so long as the tool has silktouch, even if it isn't the correct tool. (pickaxe should be preferred

Furthermore this commit adds a anti-switch whitelist, items on this list will not activate auto tool when used.
Without this using world edit with auto-tool is a pain, this fixes the issue and allows for more customization of the auto-tool module.